+++ src/PropertyManager/Travely.PropertyManager.Service/Models/Commands/AddPropertyCommand.cs
@@ -3,7 +3,7 @@
     public class AddPropertyCommand
     {
         public string Name { get; set; }
-        public float Stars { get; set; }
+        public byte Stars { get; set; }
         public string Address { get; set; }
         public double? Latitude { get; set; }
         public double? Longitude { get; set; }
